[cpp] Minor Changes. Changed static members initialization into enum to make the code build with clang, and probably visual c++, removed some dead code detected by clang static analyzer
git-svn-id: https://zxing.googlecode.com/svn/trunk@1689 59b500cc-1b3d-0410-9834-0bbf25fbcc57

+++ cpp/core/src/zxing/common/BitMatrix.cpp
@@ -91,7 +91,7 @@ void BitMatrix::clear() {
 }
 
 void BitMatrix::setRegion(size_t left, size_t top, size_t width, size_t height) {
-  if (top < 0 || left < 0) {
+  if ((long)top < 0 || (long)left < 0) {
     throw IllegalArgumentException("topI and leftJ must be nonnegative");
   }
   if (height < 1 || width < 1) {
